[feat] Visit Card 생성·조회 기능 구현

+++ src/main/java/com/walktowall/backend/visitcard/VisitCardRepository.java
@@ -2,5 +2,5 @@ package com.walktowall.backend.visitcard;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitCardRepository extends JpaRepository<VisitCard, Long> {
+public interface VisitCardRepository extends JpaRepository<VisitCard, Integer> {
 }
